test: Drop Go 1.9 specialization
Go 1.9 changed how anonymous fields were constructed via reflection
so our tests had a helper function anonymousField that supported both,
pre Go 1.9 and post Go 1.9 behaviors.

That is no longer necessary since we do not support Go 1.9.
